Create own test file for patient death
